requirements.txt
chainlit
reportlab
openai
python-dotenv